requirements.txt
git+https://github.com/kshalm/motorLib
git+https://github.com/kshalm/zmqhelpers
pyzmq
scipy>=1.1.0
pyYAML
redis
requests
loguru